fix bug in scroll animation duration
On line 80 of ‘scroll’ method, duration was overwriting the passed
value when the variable was instantiated.  Because it was null,
duration would always fall back to the default duration even if a
duration was passed.

Fix: Rename the passed duration variable name on line 77 so it doesn’t
get overwritten when ‘duration’ is instantiated in the jQuery loop on
line 80.