add safety check on scroll method
When Scroller is destroyed right after starting to animate the scroll
position, it throws an exception. (TypeError: 'undefined' is not an
object (evaluating 'data.horizontal') on line 95) Catch if the data
attribute is already gone and don’t try to do anything if that’s the
case. Tested and confirmed this will prevent the error.

The diff looks really strange because of a block indent, but the only
thing actually modified is scroller.js line 73.